Android: Show URL in OpenLinkView
